@@ -5,9 +5,10 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Pgsql;
 
 use JsonException;
-use PDO;
 use Throwable;
 use Yiisoft\Arrays\ArrayHelper;
+use Yiisoft\Db\Cache\SchemaCache;
+use Yiisoft\Db\Connection\ConnectionInterface;
 use Yiisoft\Db\Constraint\CheckConstraint;
 use Yiisoft\Db\Constraint\Constraint;
 use Yiisoft\Db\Constraint\DefaultValueConstraint;
@@ -18,8 +19,9 @@ use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\Schema\ColumnSchemaBuilder;
+use Yiisoft\Db\Schema\ColumnSchemaInterface;
 use Yiisoft\Db\Schema\Schema as AbstractSchema;
-use Yiisoft\Db\View\ViewFinderTrait;
+use Yiisoft\Db\Schema\TableSchemaInterface;
 
 use function array_change_key_case;
 use function array_merge;
@@ -27,7 +29,6 @@ use function array_unique;
 use function array_values;
 use function bindec;
 use function explode;
-use function implode;
 use function preg_match;
 use function preg_replace;
 use function str_replace;
@@ -83,14 +84,14 @@ use function substr;
  */
 final class Schema extends AbstractSchema
 {
-    use ViewFinderTrait;
-
     public const TYPE_JSONB = 'jsonb';
 
     /**
-     * @var array<array-key, string> mapping from physical column types (keys) to abstract column types (values).
+     * @var array The mapping from physical column types (keys) to abstract column types (values).
      *
      * {@see http://www.postgresql.org/docs/current/static/datatype.html#DATATYPE-TABLE}
+     *
+     * @psalm-var string[]
      */
     private array $typeMap = [
         'bit' => self::TYPE_INTEGER,
@@ -158,6 +159,11 @@ final class Schema extends AbstractSchema
         'xml' => self::TYPE_STRING,
     ];
 
+    public function __construct(private ConnectionInterface $db, SchemaCache $schemaCache)
+    {
+        parent::__construct($schemaCache);
+    }
+
     /**
      * @var string|null the default schema used for the current session.
      */
@@ -167,18 +173,18 @@ final class Schema extends AbstractSchema
      * @var string|string[] character used to quote schema, table, etc. names. An array of 2 characters can be used in
      * case starting and ending characters are different.
      */
-    protected $tableQuoteCharacter = '"';
+    protected string|array $tableQuoteCharacter = '"';
 
     /**
      * Resolves the table name and schema name (if any).
      *
      * @param string $name the table name.
      *
-     * @return TableSchema with resolved table, schema, etc. names.
+     * @return TableSchemaInterface with resolved table, schema, etc. names.
      *
-     * {@see TableSchema}
+     * {@see TableSchemaInterface}
      */
-    protected function resolveTableName(string $name): TableSchema
+    protected function resolveTableName(string $name): TableSchemaInterface
     {
         $resolvedName = new TableSchema();
 
@@ -215,14 +221,19 @@ final class Schema extends AbstractSchema
      */
     protected function findSchemaNames(): array
     {
-        $sql = <<<'SQL'
-            SELECT "ns"."nspname"
-            FROM "pg_namespace" AS "ns"
-            WHERE "ns"."nspname" != 'information_schema' AND "ns"."nspname" NOT LIKE 'pg_%'
-            ORDER BY "ns"."nspname" ASC
-            SQL;
+        $sql = <<<SQL
+        SELECT "ns"."nspname"
+        FROM "pg_namespace" AS "ns"
+        WHERE "ns"."nspname" != 'information_schema' AND "ns"."nspname" NOT LIKE 'pg_%'
+        ORDER BY "ns"."nspname" ASC
+        SQL;
 
-        return $this->getDb()->createCommand($sql)->queryColumn();
+        $schemaNames = $this->db->createCommand($sql)->queryColumn();
+        if (!$schemaNames) {
+            return [];
+        }
+
+        return $schemaNames;
     }
 
     /**
@@ -243,15 +254,20 @@ final class Schema extends AbstractSchema
             $schema = $this->defaultSchema;
         }
 
-        $sql = <<<'SQL'
-            SELECT c.relname AS table_name
-            FROM pg_class c
-            INNER JOIN pg_namespace ns ON ns.oid = c.relnamespace
-            WHERE ns.nspname = :schemaName AND c.relkind IN ('r','v','m','f', 'p')
-            ORDER BY c.relname
-            SQL;
+        $sql = <<<SQL
+        SELECT c.relname AS table_name
+        FROM pg_class c
+        INNER JOIN pg_namespace ns ON ns.oid = c.relnamespace
+        WHERE ns.nspname = :schemaName AND c.relkind IN ('r','v','m','f', 'p')
+        ORDER BY c.relname
+        SQL;
 
-        return $this->getDb()->createCommand($sql, [':schemaName' => $schema])->queryColumn();
+        $tableNames = $this->db->createCommand($sql, [':schemaName' => $schema])->queryColumn();
+        if (!$tableNames) {
+            return [];
+        }
+
+        return $tableNames;
     }
 
     /**
@@ -259,11 +275,11 @@ final class Schema extends AbstractSchema
      *
      * @param string $name table name.
      *
-     * @throws Exception|InvalidConfigException
+     * @throws Exception|InvalidConfigException|Throwable
      *
-     * @return TableSchema|null DBMS-dependent table metadata, `null` if the table does not exist.
+     * @return TableSchemaInterface|null DBMS-dependent table metadata, `null` if the table does not exist.
      */
-    protected function loadTableSchema(string $name): ?TableSchema
+    protected function loadTableSchema(string $name): ?TableSchemaInterface
     {
         $table = new TableSchema();
 
@@ -282,13 +298,13 @@ final class Schema extends AbstractSchema
      *
      * @param string $tableName table name.
      *
-     * @throws Exception|InvalidConfigException
+     * @throws Exception|InvalidConfigException|Throwable
      *
      * @return Constraint|null primary key for the given table, `null` if the table has no primary key.
      */
     protected function loadTablePrimaryKey(string $tableName): ?Constraint
     {
-        $tablePrimaryKey = $this->loadTableConstraints($tableName, 'primaryKey');
+        $tablePrimaryKey = $this->loadTableConstraints($tableName, self::PRIMARY_KEY);
 
         return $tablePrimaryKey instanceof Constraint ? $tablePrimaryKey : null;
     }
@@ -298,13 +314,15 @@ final class Schema extends AbstractSchema
      *
      * @param string $tableName table name.
      *
-     * @throws Exception|InvalidConfigException
+     * @throws Exception|InvalidConfigException|Throwable
      *
-     * @return array|ForeignKeyConstraint[] foreign keys for the given table.
+     * @return array foreign keys for the given table.
+     *
+     * @psaml-return array|ForeignKeyConstraint[]
      */
     protected function loadTableForeignKeys(string $tableName): array
     {
-        $tableForeignKeys = $this->loadTableConstraints($tableName, 'foreignKeys');
+        $tableForeignKeys = $this->loadTableConstraints($tableName, self::FOREIGN_KEYS);
 
         return is_array($tableForeignKeys) ? $tableForeignKeys : [];
     }
@@ -320,34 +338,34 @@ final class Schema extends AbstractSchema
      */
     protected function loadTableIndexes(string $tableName): array
     {
-        $sql = <<<'SQL'
-            SELECT
-                "ic"."relname" AS "name",
-                "ia"."attname" AS "column_name",
-                "i"."indisunique" AS "index_is_unique",
-                "i"."indisprimary" AS "index_is_primary"
-            FROM "pg_class" AS "tc"
-            INNER JOIN "pg_namespace" AS "tcns"
-                ON "tcns"."oid" = "tc"."relnamespace"
-            INNER JOIN "pg_index" AS "i"
-                ON "i"."indrelid" = "tc"."oid"
-            INNER JOIN "pg_class" AS "ic"
-                ON "ic"."oid" = "i"."indexrelid"
-            INNER JOIN "pg_attribute" AS "ia"
-                ON "ia"."attrelid" = "i"."indexrelid"
-            WHERE "tcns"."nspname" = :schemaName AND "tc"."relname" = :tableName
-            ORDER BY "ia"."attnum" ASC
-            SQL;
+        $sql = <<<SQL
+        SELECT
+            "ic"."relname" AS "name",
+            "ia"."attname" AS "column_name",
+            "i"."indisunique" AS "index_is_unique",
+            "i"."indisprimary" AS "index_is_primary"
+        FROM "pg_class" AS "tc"
+        INNER JOIN "pg_namespace" AS "tcns"
+            ON "tcns"."oid" = "tc"."relnamespace"
+        INNER JOIN "pg_index" AS "i"
+            ON "i"."indrelid" = "tc"."oid"
+        INNER JOIN "pg_class" AS "ic"
+            ON "ic"."oid" = "i"."indexrelid"
+        INNER JOIN "pg_attribute" AS "ia"
+            ON "ia"."attrelid" = "i"."indexrelid"
+        WHERE "tcns"."nspname" = :schemaName AND "tc"."relname" = :tableName
+        ORDER BY "ia"."attnum" ASC
+        SQL;
 
         $resolvedName = $this->resolveTableName($tableName);
 
-        $indexes = $this->getDb()->createCommand($sql, [
+        $indexes = $this->db->createCommand($sql, [
             ':schemaName' => $resolvedName->getSchemaName(),
             ':tableName' => $resolvedName->getName(),
         ])->queryAll();
 
-        /** @var array<array-key, array<array-key, mixed>> @indexes */
-        $indexes = $this->normalizePdoRowKeyCase($indexes, true);
+        /** @var array[] @indexes */
+        $indexes = $this->normalizeRowKeyCase($indexes, true);
         $indexes = ArrayHelper::index($indexes, null, 'name');
         $result = [];
 
@@ -381,13 +399,15 @@ final class Schema extends AbstractSchema
      *
      * @param string $tableName table name.
      *
-     * @throws Exception|InvalidConfigException
+     * @throws Exception|InvalidConfigException|Throwable
      *
-     * @return array|Constraint[] unique constraints for the given table.
+     * @return array unique constraints for the given table.
+     *
+     * @psalm-return array|Constraint[]
      */
     protected function loadTableUniques(string $tableName): array
     {
-        $tableUniques = $this->loadTableConstraints($tableName, 'uniques');
+        $tableUniques = $this->loadTableConstraints($tableName, self::UNIQUES);
 
         return is_array($tableUniques) ? $tableUniques : [];
     }
@@ -397,13 +417,15 @@ final class Schema extends AbstractSchema
      *
      * @param string $tableName table name.
      *
-     * @throws Exception|InvalidConfigException
+     * @throws Exception|InvalidConfigException|Throwable
      *
-     * @return array|CheckConstraint[] check constraints for the given table.
+     * @return array check constraints for the given table.
+     *
+     * @psaml-return array|CheckConstraint[]
      */
     protected function loadTableChecks(string $tableName): array
     {
-        $tableChecks = $this->loadTableConstraints($tableName, 'checks');
+        $tableChecks = $this->loadTableConstraints($tableName, self::CHECKS);
 
         return is_array($tableChecks) ? $tableChecks : [];
     }
@@ -423,22 +445,12 @@ final class Schema extends AbstractSchema
     }
 
     /**
-     * Creates a query builder for the PostgreSQL database.
-     *
-     * @return QueryBuilder query builder instance
-     */
-    public function createQueryBuilder(): QueryBuilder
-    {
-        return new QueryBuilder($this->getDb());
-    }
-
-    /**
      * Resolves the table name and schema name (if any).
      *
-     * @param TableSchema $table the table metadata object.
+     * @param TableSchemaInterface $table the table metadata object.
      * @param string $name the table name
      */
-    protected function resolveTableNames(TableSchema $table, string $name): void
+    protected function resolveTableNames(TableSchemaInterface $table, string $name): void
     {
         $parts = explode('.', str_replace('"', '', $name));
 
@@ -459,39 +471,49 @@ final class Schema extends AbstractSchema
         $table->fullName($name);
     }
 
+    /**
+     * @throws Exception|InvalidConfigException|Throwable
+     */
     protected function findViewNames(string $schema = ''): array
     {
         if ($schema === '') {
             $schema = $this->defaultSchema;
         }
 
-        $sql = <<<'SQL'
-            SELECT c.relname AS table_name
-            FROM pg_class c
-            INNER JOIN pg_namespace ns ON ns.oid = c.relnamespace
-            WHERE ns.nspname = :schemaName AND (c.relkind = 'v' OR c.relkind = 'm')
-            ORDER BY c.relname
-            SQL;
+        $sql = <<<SQL
+        SELECT c.relname AS table_name
+        FROM pg_class c
+        INNER JOIN pg_namespace ns ON ns.oid = c.relnamespace
+        WHERE ns.nspname = :schemaName AND (c.relkind = 'v' OR c.relkind = 'm')
+        ORDER BY c.relname
+        SQL;
 
-        return $this->getDb()->createCommand($sql, [':schemaName' => $schema])->queryColumn();
+        $viewNames = $this->db->createCommand($sql, [':schemaName' => $schema])->queryColumn();
+        if (!$viewNames) {
+            return [];
+        }
+
+        return $viewNames;
     }
 
     /**
      * Collects the foreign key column details for the given table.
      *
-     * @param TableSchema $table the table metadata
+     * @param TableSchemaInterface $table the table metadata
      *
      * @throws Exception|InvalidConfigException|Throwable
      */
-    protected function findConstraints(TableSchema $table): void
+    protected function findConstraints(TableSchemaInterface $table): void
     {
         $tableName = $table->getName();
         $tableSchema = $table->getSchemaName();
 
-        $tableName = $this->quoteValue($tableName);
+        /** @var mixed */
+        $tableName = $this->db->getQuoter()->quoteValue($tableName);
 
         if ($tableSchema !== null) {
-            $tableSchema = $this->quoteValue($tableSchema);
+            /** @var mixed */
+            $tableSchema = $this->db->getQuoter()->quoteValue($tableSchema);
         }
 
         /**
@@ -500,40 +522,56 @@ final class Schema extends AbstractSchema
          */
 
         $sql = <<<SQL
-            SELECT
-                ct.conname as constraint_name,
-                a.attname as column_name,
-                fc.relname as foreign_table_name,
-                fns.nspname as foreign_table_schema,
-                fa.attname as foreign_column_name
+        SELECT
+            ct.conname as constraint_name,
+            a.attname as column_name,
+            fc.relname as foreign_table_name,
+            fns.nspname as foreign_table_schema,
+            fa.attname as foreign_column_name
             FROM
-                (SELECT ct.conname, ct.conrelid, ct.confrelid, ct.conkey, ct.contype, ct.confkey,
-                        generate_subscripts(ct.conkey, 1) AS s
-                   FROM pg_constraint ct
-                ) AS ct
-                inner join pg_class c on c.oid=ct.conrelid
-                inner join pg_namespace ns on c.relnamespace=ns.oid
-                inner join pg_attribute a on a.attrelid=ct.conrelid and a.attnum = ct.conkey[ct.s]
-                left join pg_class fc on fc.oid=ct.confrelid
-                left join pg_namespace fns on fc.relnamespace=fns.oid
-                left join pg_attribute fa on fa.attrelid=ct.confrelid and fa.attnum = ct.confkey[ct.s]
-            WHERE
-                ct.contype='f'
-                and c.relname={$tableName}
-                and ns.nspname={$tableSchema}
-            ORDER BY
-                fns.nspname, fc.relname, a.attnum
-            SQL;
+            (SELECT ct.conname, ct.conrelid, ct.confrelid, ct.conkey, ct.contype, ct.confkey,
+                generate_subscripts(ct.conkey, 1) AS s
+                FROM pg_constraint ct
+            ) AS ct
+            inner join pg_class c on c.oid=ct.conrelid
+            inner join pg_namespace ns on c.relnamespace=ns.oid
+            inner join pg_attribute a on a.attrelid=ct.conrelid and a.attnum = ct.conkey[ct.s]
+            left join pg_class fc on fc.oid=ct.confrelid
+            left join pg_namespace fns on fc.relnamespace=fns.oid
+            left join pg_attribute fa on fa.attrelid=ct.confrelid and fa.attnum = ct.confkey[ct.s]
+        WHERE
+            ct.contype='f'
+            and c.relname=$tableName
+            and ns.nspname=$tableSchema
+        ORDER BY
+            fns.nspname, fc.relname, a.attnum
+        SQL;
 
         /** @var array{array{tableName: string, columns: array}} $constraints */
         $constraints = [];
-        $slavePdo = $this->getDb()->getSlavePdo();
 
-        /** @var FindConstraintArray $constraint */
-        foreach ($this->getDb()->createCommand($sql)->queryAll() as $constraint) {
-            if ($slavePdo !== null && $slavePdo->getAttribute(PDO::ATTR_CASE) === PDO::CASE_UPPER) {
-                $constraint = array_change_key_case($constraint, CASE_LOWER);
-            }
+        /**
+         * @psalm-var array<
+         *   array{
+         *     constraint_name: string,
+         *     column_name: string,
+         *     foreign_table_name: string,
+         *     foreign_table_schema: string,
+         *     foreign_column_name: string,
+         *   }
+         * > $rows
+         */
+        $rows = $this->db->createCommand($sql)->queryAll();
+
+        foreach ($rows as $constraint) {
+            /** @psalm-var array{
+             *     constraint_name: string,
+             *     column_name: string,
+             *     foreign_table_name: string,
+             *     foreign_table_schema: string,
+             *     foreign_column_name: string,
+             *   } $constraint */
+            $constraint = $this->normalizeRowKeyCase($constraint, false);
 
             if ($constraint['foreign_table_schema'] !== $this->defaultSchema) {
                 $foreignTable = $constraint['foreign_table_schema'] . '.' . $constraint['foreign_table_name'];
@@ -568,31 +606,31 @@ final class Schema extends AbstractSchema
     /**
      * Gets information about given table unique indexes.
      *
-     * @param TableSchema $table the table metadata.
+     * @param TableSchemaInterface $table the table metadata.
      *
      * @throws Exception|InvalidConfigException|Throwable
      *
      * @return array with index and column names.
      */
-    protected function getUniqueIndexInformation(TableSchema $table): array
+    protected function getUniqueIndexInformation(TableSchemaInterface $table): array
     {
         $sql = <<<'SQL'
-            SELECT
-                i.relname as indexname,
-                pg_get_indexdef(idx.indexrelid, k + 1, TRUE) AS columnname
-            FROM (
-              SELECT *, generate_subscripts(indkey, 1) AS k
-              FROM pg_index
-            ) idx
-            INNER JOIN pg_class i ON i.oid = idx.indexrelid
-            INNER JOIN pg_class c ON c.oid = idx.indrelid
-            INNER JOIN pg_namespace ns ON c.relnamespace = ns.oid
-            WHERE idx.indisprimary = FALSE AND idx.indisunique = TRUE
-            AND c.relname = :tableName AND ns.nspname = :schemaName
-            ORDER BY i.relname, k
-            SQL;
+        SELECT
+            i.relname as indexname,
+            pg_get_indexdef(idx.indexrelid, k + 1, TRUE) AS columnname
+        FROM (
+            SELECT *, generate_subscripts(indkey, 1) AS k
+            FROM pg_index
+        ) idx
+        INNER JOIN pg_class i ON i.oid = idx.indexrelid
+        INNER JOIN pg_class c ON c.oid = idx.indrelid
+        INNER JOIN pg_namespace ns ON c.relnamespace = ns.oid
+        WHERE idx.indisprimary = FALSE AND idx.indisunique = TRUE
+        AND c.relname = :tableName AND ns.nspname = :schemaName
+        ORDER BY i.relname, k
+        SQL;
 
-        return $this->getDb()->createCommand($sql, [
+        return $this->db->createCommand($sql, [
             ':schemaName' => $table->getSchemaName(),
             ':tableName' => $table->getName(),
         ])->queryAll();
@@ -610,22 +648,20 @@ final class Schema extends AbstractSchema
      * ]
      * ```
      *
-     * @param TableSchema $table the table metadata
+     * @param TableSchemaInterface $table the table metadata
      *
      * @throws Exception|InvalidConfigException|Throwable
      *
      * @return array all unique indexes for the given table.
      */
-    public function findUniqueIndexes(TableSchema $table): array
+    public function findUniqueIndexes(TableSchemaInterface $table): array
     {
         $uniqueIndexes = [];
-        $slavePdo = $this->getDb()->getSlavePdo();
 
         /** @var array{indexname: string, columnname: string} $row */
         foreach ($this->getUniqueIndexInformation($table) as $row) {
-            if ($slavePdo !== null && $slavePdo->getAttribute(PDO::ATTR_CASE) === PDO::CASE_UPPER) {
-                $row = array_change_key_case($row, CASE_LOWER);
-            }
+            /** @var array{indexname: string, columnname: string} $row */
+            $row = $this->normalizeRowKeyCase($row, false);
 
             $column = $row['columnname'];
 
@@ -647,83 +683,85 @@ final class Schema extends AbstractSchema
     /**
      * Collects the metadata of table columns.
      *
-     * @param TableSchema $table the table metadata.
+     * @param TableSchemaInterface $table the table metadata.
      *
      * @throws Exception|InvalidConfigException|JsonException|Throwable
      *
      * @return bool whether the table exists in the database.
      */
-    protected function findColumns(TableSchema $table): bool
+    protected function findColumns(TableSchemaInterface $table): bool
     {
         $tableName = $table->getName();
         $schemaName = $table->getSchemaName();
         $orIdentity = '';
 
-        $tableName = $this->getDb()->quoteValue($tableName);
+        /** @var mixed */
+        $tableName = $this->db->getQuoter()->quoteValue($tableName);
 
         if ($schemaName !== null) {
-            $schemaName = $this->getDb()->quoteValue($schemaName);
+            /** @var mixed */
+            $schemaName = $this->db->getQuoter()->quoteValue($schemaName);
         }
 
-        if (version_compare($this->getDb()->getServerVersion(), '12.0', '>=')) {
+        if (version_compare($this->db->getServerVersion(), '12.0', '>=')) {
             $orIdentity = 'OR a.attidentity != \'\'';
         }
 
         $sql = <<<SQL
-            SELECT
-                d.nspname AS table_schema,
-                c.relname AS table_name,
-                a.attname AS column_name,
-                COALESCE(td.typname, tb.typname, t.typname) AS data_type,
-                COALESCE(td.typtype, tb.typtype, t.typtype) AS type_type,
-                a.attlen AS character_maximum_length,
-                pg_catalog.col_description(c.oid, a.attnum) AS column_comment,
-                a.atttypmod AS modifier,
-                a.attnotnull = false AS is_nullable,
-                CAST(pg_get_expr(ad.adbin, ad.adrelid) AS varchar) AS column_default,
-                coalesce(pg_get_expr(ad.adbin, ad.adrelid) ~ 'nextval',false) {$orIdentity} AS is_autoinc,
-                pg_get_serial_sequence(quote_ident(d.nspname) || '.' || quote_ident(c.relname), a.attname)
-                    AS sequence_name,
-                CASE WHEN COALESCE(td.typtype, tb.typtype, t.typtype) = 'e'::char
-                    THEN array_to_string(
-                        (
-                            SELECT array_agg(enumlabel)
-                            FROM pg_enum
-                            WHERE enumtypid = COALESCE(td.oid, tb.oid, a.atttypid)
-                            )::varchar[],
-                        ',')
-                    ELSE NULL
-                END AS enum_values,
-                CASE atttypid
-                     WHEN 21 /*int2*/ THEN 16
-                     WHEN 23 /*int4*/ THEN 32
-                     WHEN 20 /*int8*/ THEN 64
-                     WHEN 1700 /*numeric*/ THEN
-                          CASE WHEN atttypmod = -1
-                           THEN null
-                           ELSE ((atttypmod - 4) >> 16) & 65535
-                           END
-                     WHEN 700 /*float4*/ THEN 24 /*FLT_MANT_DIG*/
-                     WHEN 701 /*float8*/ THEN 53 /*DBL_MANT_DIG*/
-                     ELSE null
-                  END   AS numeric_precision,
-                  CASE
-                    WHEN atttypid IN (21, 23, 20) THEN 0
-                    WHEN atttypid IN (1700) THEN
-                    CASE
-                        WHEN atttypmod = -1 THEN null
-                        ELSE (atttypmod - 4) & 65535
+        SELECT
+            d.nspname AS table_schema,
+            c.relname AS table_name,
+            a.attname AS column_name,
+            COALESCE(td.typname, tb.typname, t.typname) AS data_type,
+            COALESCE(td.typtype, tb.typtype, t.typtype) AS type_type,
+            a.attlen AS character_maximum_length,
+            pg_catalog.col_description(c.oid, a.attnum) AS column_comment,
+            a.atttypmod AS modifier,
+            a.attnotnull = false AS is_nullable,
+            CAST(pg_get_expr(ad.adbin, ad.adrelid) AS varchar) AS column_default,
+            coalesce(pg_get_expr(ad.adbin, ad.adrelid) ~ 'nextval',false) $orIdentity AS is_autoinc,
+            pg_get_serial_sequence(quote_ident(d.nspname) || '.' || quote_ident(c.relname), a.attname)
+            AS sequence_name,
+            CASE WHEN COALESCE(td.typtype, tb.typtype, t.typtype) = 'e'::char
+                THEN array_to_string(
+                    (
+                        SELECT array_agg(enumlabel)
+                        FROM pg_enum
+                        WHERE enumtypid = COALESCE(td.oid, tb.oid, a.atttypid)
+                    )::varchar[],
+                ',')
+                ELSE NULL
+            END AS enum_values,
+            CASE atttypid
+                WHEN 21 /*int2*/ THEN 16
+                WHEN 23 /*int4*/ THEN 32
+                WHEN 20 /*int8*/ THEN 64
+                WHEN 1700 /*numeric*/ THEN
+                    CASE WHEN atttypmod = -1
+                        THEN null
+                        ELSE ((atttypmod - 4) >> 16) & 65535
+                        END
+                WHEN 700 /*float4*/ THEN 24 /*FLT_MANT_DIG*/
+                WHEN 701 /*float8*/ THEN 53 /*DBL_MANT_DIG*/
+                    ELSE null
+                    END   AS numeric_precision,
+            CASE
+                WHEN atttypid IN (21, 23, 20) THEN 0
+                WHEN atttypid IN (1700) THEN
+            CASE
+                WHEN atttypmod = -1 THEN null
+                    ELSE (atttypmod - 4) & 65535
                     END
-                       ELSE null
-                  END AS numeric_scale,
-                CAST(
-                         information_schema._pg_char_max_length(
-                             information_schema._pg_truetypid(a, t),
-                             information_schema._pg_truetypmod(a, t)
-                             ) AS numeric
-                ) AS size,
-                a.attnum = any (ct.conkey) as is_pkey,
-                COALESCE(NULLIF(a.attndims, 0), NULLIF(t.typndims, 0), (t.typcategory='A')::int) AS dimension
+                    ELSE null
+                    END AS numeric_scale,
+                    CAST(
+                        information_schema._pg_char_max_length(
+                        information_schema._pg_truetypid(a, t),
+                        information_schema._pg_truetypmod(a, t)
+                        ) AS numeric
+                    ) AS size,
+                    a.attnum = any (ct.conkey) as is_pkey,
+                    COALESCE(NULLIF(a.attndims, 0), NULLIF(t.typndims, 0), (t.typcategory='A')::int) AS dimension
             FROM
                 pg_class c
                 LEFT JOIN pg_attribute a ON a.attrelid = c.oid
@@ -736,31 +774,29 @@ final class Schema extends AbstractSchema
                 LEFT JOIN pg_constraint ct ON ct.conrelid = c.oid AND ct.contype = 'p'
             WHERE
                 a.attnum > 0 AND t.typname != '' AND NOT a.attisdropped
-                AND c.relname = {$tableName}
-                AND d.nspname = {$schemaName}
+                AND c.relname = $tableName
+                AND d.nspname = $schemaName
             ORDER BY
                 a.attnum;
-            SQL;
+        SQL;
 
-        /** @var array columns */
-        $columns = $this->getDb()->createCommand($sql)->queryAll();
-        $slavePdo = $this->getDb()->getSlavePdo();
+        $columns = $this->db->createCommand($sql)->queryAll();
 
         if (empty($columns)) {
             return false;
         }
 
-        /** @var array<array-key, mixed> $column */
+        /** @var array $column */
         foreach ($columns as $column) {
-            if ($slavePdo !== null && $slavePdo->getAttribute(PDO::ATTR_CASE) === PDO::CASE_UPPER) {
-                $column = array_change_key_case($column, CASE_LOWER);
-            }
-
             /** @psalm-var ColumnArray $column */
+            $column = $this->normalizeRowKeyCase($column, false);
+
+            /** @psalm-var ColumnSchema $loadColumnSchema */
             $loadColumnSchema = $this->loadColumnSchema($column);
+
             $table->columns($loadColumnSchema->getName(), $loadColumnSchema);
 
-            /** @var mixed $defaultValue */
+            /** @var mixed */
             $defaultValue = $loadColumnSchema->getDefaultValue();
 
             if ($loadColumnSchema->isPrimaryKey()) {
@@ -813,9 +849,9 @@ final class Schema extends AbstractSchema
     }
 
     /**
-     * Loads the column information into a {@see ColumnSchema} object.
+     * Loads the column information into a {@see ColumnSchemaInterface} object.
      *
-     * @param array{
+     * @psalm-param array{
      *   table_schema: string,
      *   table_name: string,
      *   column_name: string,
@@ -836,9 +872,9 @@ final class Schema extends AbstractSchema
      *   dimension: int
      * } $info column information.
      *
-     * @return ColumnSchema the column schema object.
+     * @return ColumnSchemaInterface the column schema object.
      */
-    protected function loadColumnSchema(array $info): ColumnSchema
+    protected function loadColumnSchema(array $info): ColumnSchemaInterface
     {
         $column = $this->createColumnSchema();
         $column->allowNull($info['is_nullable']);
@@ -891,45 +927,6 @@ final class Schema extends AbstractSchema
     }
 
     /**
-     * Executes the INSERT command, returning primary key values.
-     *
-     * @param string $table the table that new rows will be inserted into.
-     * @param array $columns the column data (name => value) to be inserted into the table.
-     *
-     * @throws Exception|InvalidConfigException|Throwable
-     *
-     * @return array|false primary key values or false if the command fails.
-     */
-    public function insert(string $table, array $columns)
-    {
-        $params = [];
-        $returnColumns = [];
-        $sql = $this->getDb()->getQueryBuilder()->insert($table, $columns, $params);
-        $tableSchema = $this->getTableSchema($table);
-
-        if ($tableSchema !== null) {
-            $returnColumns = $tableSchema->getPrimaryKey();
-        }
-
-        if (!empty($returnColumns)) {
-            $returning = [];
-            /** @var string $name */
-            foreach ($returnColumns as $name) {
-                $returning[] = $this->quoteColumnName($name);
-            }
-            $sql .= ' RETURNING ' . implode(', ', $returning);
-        }
-
-        $command = $this->getDb()->createCommand($sql, $params);
-        $command->prepare(false);
-        $result = $command->queryOne();
-
-        $pdoStatement = $command->getPdoStatement();
-
-        return $pdoStatement !== null && !$pdoStatement->rowCount() ? false : $result;
-    }
-
-    /**
      * Loads multiple types of constraints and returns the specified ones.
      *
      * @param string $tableName table name.
@@ -941,40 +938,37 @@ final class Schema extends AbstractSchema
      *
      * @throws Exception|InvalidConfigException|Throwable
      *
-     * @return (CheckConstraint|Constraint|ForeignKeyConstraint)[]|Constraint|null constraints.
-     *
-     * @psalm-return Constraint|list<CheckConstraint|Constraint|ForeignKeyConstraint>|null
+     * @return array|Constraint|null (CheckConstraint|Constraint|ForeignKeyConstraint)[]|Constraint|null constraints.
      */
-    private function loadTableConstraints(string $tableName, string $returnType)
+    private function loadTableConstraints(string $tableName, string $returnType): array|Constraint|null
     {
-        /** @var string $sql */
-        $sql = <<<'SQL'
-            SELECT
-                "c"."conname" AS "name",
-                "a"."attname" AS "column_name",
-                "c"."contype" AS "type",
-                "ftcns"."nspname" AS "foreign_table_schema",
-                "ftc"."relname" AS "foreign_table_name",
-                "fa"."attname" AS "foreign_column_name",
-                "c"."confupdtype" AS "on_update",
-                "c"."confdeltype" AS "on_delete",
-                pg_get_constraintdef("c"."oid") AS "check_expr"
-            FROM "pg_class" AS "tc"
-            INNER JOIN "pg_namespace" AS "tcns"
-                ON "tcns"."oid" = "tc"."relnamespace"
-            INNER JOIN "pg_constraint" AS "c"
-                ON "c"."conrelid" = "tc"."oid"
-            INNER JOIN "pg_attribute" AS "a"
-                ON "a"."attrelid" = "c"."conrelid" AND "a"."attnum" = ANY ("c"."conkey")
-            LEFT JOIN "pg_class" AS "ftc"
-                ON "ftc"."oid" = "c"."confrelid"
-            LEFT JOIN "pg_namespace" AS "ftcns"
-                ON "ftcns"."oid" = "ftc"."relnamespace"
-            LEFT JOIN "pg_attribute" "fa"
-                ON "fa"."attrelid" = "c"."confrelid" AND "fa"."attnum" = ANY ("c"."confkey")
-            WHERE "tcns"."nspname" = :schemaName AND "tc"."relname" = :tableName
-            ORDER BY "a"."attnum" ASC, "fa"."attnum" ASC
-            SQL;
+        $sql = <<<SQL
+        SELECT
+            "c"."conname" AS "name",
+            "a"."attname" AS "column_name",
+            "c"."contype" AS "type",
+            "ftcns"."nspname" AS "foreign_table_schema",
+            "ftc"."relname" AS "foreign_table_name",
+            "fa"."attname" AS "foreign_column_name",
+            "c"."confupdtype" AS "on_update",
+            "c"."confdeltype" AS "on_delete",
+            pg_get_constraintdef("c"."oid") AS "check_expr"
+        FROM "pg_class" AS "tc"
+        INNER JOIN "pg_namespace" AS "tcns"
+            ON "tcns"."oid" = "tc"."relnamespace"
+        INNER JOIN "pg_constraint" AS "c"
+            ON "c"."conrelid" = "tc"."oid"
+        INNER JOIN "pg_attribute" AS "a"
+            ON "a"."attrelid" = "c"."conrelid" AND "a"."attnum" = ANY ("c"."conkey")
+        LEFT JOIN "pg_class" AS "ftc"
+            ON "ftc"."oid" = "c"."confrelid"
+        LEFT JOIN "pg_namespace" AS "ftcns"
+            ON "ftcns"."oid" = "ftc"."relnamespace"
+        LEFT JOIN "pg_attribute" "fa"
+            ON "fa"."attrelid" = "c"."confrelid" AND "fa"."attnum" = ANY ("c"."confkey")
+        WHERE "tcns"."nspname" = :schemaName AND "tc"."relname" = :tableName
+        ORDER BY "a"."attnum" ASC, "fa"."attnum" ASC
+        SQL;
 
         /** @var array<array-key, string> $actionTypes */
         $actionTypes = [
@@ -987,20 +981,20 @@ final class Schema extends AbstractSchema
 
         $resolvedName = $this->resolveTableName($tableName);
 
-        $constraints = $this->getDb()->createCommand($sql, [
+        $constraints = $this->db->createCommand($sql, [
             ':schemaName' => $resolvedName->getSchemaName(),
             ':tableName' => $resolvedName->getName(),
         ])->queryAll();
 
         /** @var array<array-key, array> $constraints */
-        $constraints = $this->normalizePdoRowKeyCase($constraints, true);
+        $constraints = $this->normalizeRowKeyCase($constraints, true);
         $constraints = ArrayHelper::index($constraints, null, ['type', 'name']);
 
         $result = [
-            'primaryKey' => null,
-            'foreignKeys' => [],
-            'uniques' => [],
-            'checks' => [],
+            self::PRIMARY_KEY => null,
+            self::FOREIGN_KEYS => [],
+            self::UNIQUES => [],
+            self::CHECKS => [],
         ];
 
         /**
@@ -1015,17 +1009,15 @@ final class Schema extends AbstractSchema
             foreach ($names as $name => $constraint) {
                 switch ($type) {
                     case 'p':
-                        $ct = (new Constraint())
+                        $result[self::PRIMARY_KEY] = (new Constraint())
                             ->name($name)
                             ->columnNames(ArrayHelper::getColumn($constraint, 'column_name'));
-
-                        $result['primaryKey'] = $ct;
                         break;
                     case 'f':
                         $onDelete = $actionTypes[$constraint[0]['on_delete']] ?? null;
                         $onUpdate = $actionTypes[$constraint[0]['on_update']] ?? null;
 
-                        $fk = (new ForeignKeyConstraint())
+                        $result[self::FOREIGN_KEYS][] = (new ForeignKeyConstraint())
                             ->name($name)
                             ->columnNames(array_values(
                                 array_unique(ArrayHelper::getColumn($constraint, 'column_name'))
@@ -1037,23 +1029,17 @@ final class Schema extends AbstractSchema
                             ))
                             ->onDelete($onDelete)
                             ->onUpdate($onUpdate);
-
-                        $result['foreignKeys'][] = $fk;
                         break;
                     case 'u':
-                        $ct = (new Constraint())
+                        $result[self::UNIQUES][] = (new Constraint())
                             ->name($name)
                             ->columnNames(ArrayHelper::getColumn($constraint, 'column_name'));
-
-                        $result['uniques'][] = $ct;
                         break;
                     case 'c':
-                        $ck = (new CheckConstraint())
+                        $result[self::CHECKS][] = (new CheckConstraint())
                             ->name($name)
                             ->columnNames(ArrayHelper::getColumn($constraint, 'column_name'))
                             ->expression($constraint[0]['check_expr']);
-
-                        $result['checks'][] = $ck;
                         break;
                 }
             }
@@ -1087,9 +1073,91 @@ final class Schema extends AbstractSchema
      * @param array|int|string|null $length length or precision of the column. See {@see ColumnSchemaBuilder::$length}.
      *
      * @return ColumnSchemaBuilder column schema builder instance
+     *
+     * @psalm-param int|string|string[]|null $length
      */
-    public function createColumnSchemaBuilder(string $type, $length = null): ColumnSchemaBuilder
+    public function createColumnSchemaBuilder(string $type, int|string|array|null $length = null): ColumnSchemaBuilder
     {
         return new ColumnSchemaBuilder($type, $length);
+    }
+
+    /**
+     * Returns the actual name of a given table name.
+     *
+     * This method will strip off curly brackets from the given table name and replace the percentage character '%' with
+     * {@see ConnectionInterface::tablePrefix}.
+     *
+     * @param string $name the table name to be converted.
+     *
+     * @return string the real name of the given table name.
+     */
+    public function getRawTableName(string $name): string
+    {
+        if (str_contains($name, '{{')) {
+            $name = preg_replace('/{{(.*?)}}/', '\1', $name);
+
+            return str_replace('%', $this->db->getTablePrefix(), $name);
+        }
+
+        return $name;
+    }
+
+    /**
+     * Returns the cache key for the specified table name.
+     *
+     * @param string $name the table name.
+     *
+     * @return array the cache key.
+     */
+    protected function getCacheKey(string $name): array
+    {
+        return array_merge([__CLASS__], $this->db->getCacheKey(), [$this->getRawTableName($name)]);
+    }
+
+    /**
+     * Returns the cache tag name.
+     *
+     * This allows {@see refresh()} to invalidate all cached table schemas.
+     *
+     * @return string the cache tag name.
+     */
+    protected function getCacheTag(): string
+    {
+        return md5(serialize(array_merge([__CLASS__], $this->db->getCacheKey())));
+    }
+
+    /**
+     * @return bool whether this DBMS supports [savepoint](http://en.wikipedia.org/wiki/Savepoint).
+     */
+    public function supportsSavepoint(): bool
+    {
+        return $this->db->isSavepointEnabled();
+    }
+
+    /**
+     * Changes row's array key case to lower.
+     *
+     * @param array $row row's array or an array of row's arrays.
+     * @param bool $multiple whether multiple rows or a single row passed.
+     *
+     * @return array normalized row or rows.
+     */
+    protected function normalizeRowKeyCase(array $row, bool $multiple): array
+    {
+        if ($multiple) {
+            return array_map(static function (array $row) {
+                return array_change_key_case($row, CASE_LOWER);
+            }, $row);
+        }
+
+        return array_change_key_case($row, CASE_LOWER);
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function getLastInsertID(?string $sequenceName = null): string
+    {
+        return $this->db->getLastInsertID($sequenceName);
     }
 }
