@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Pgsql;
+namespace Yiisoft\Db\Pgsql\Builder;
 
 use JsonException;
 use Yiisoft\Db\Exception\Exception;
@@ -11,10 +11,9 @@ use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Expression\ArrayExpression;
 use Yiisoft\Db\Expression\ExpressionBuilderInterface;
-use Yiisoft\Db\Expression\ExpressionBuilderTrait;
 use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Expression\JsonExpression;
-use Yiisoft\Db\Query\Query;
+use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 use Yiisoft\Db\Query\QueryInterface;
 use Yiisoft\Json\Json;
 
@@ -23,7 +22,9 @@ use Yiisoft\Json\Json;
  */
 final class JsonExpressionBuilder implements ExpressionBuilderInterface
 {
-    use ExpressionBuilderTrait;
+    public function __construct(private QueryBuilderInterface $queryBuilder)
+    {
+    }
 
     /**
      * Method builds the raw SQL from the $expression that will not be additionally escaped or quoted.
@@ -43,7 +44,7 @@ final class JsonExpressionBuilder implements ExpressionBuilderInterface
          */
         $value = $expression->getValue();
 
-        if ($value instanceof Query) {
+        if ($value instanceof QueryInterface) {
             [$sql, $params] = $this->queryBuilder->build($value, $params);
             return "($sql)" . $this->getTypecast($expression);
         }
