@@ -6,11 +6,10 @@ namespace Yiisoft\Db\Pgsql\Tests;
 
 use PDO;
 use Yiisoft\Cache\CacheKeyNormalizer;
-use Yiisoft\Db\Exception\Exception;
+use Yiisoft\Db\Connection\ConnectionInterface;
 use Yiisoft\Db\Exception\InvalidConfigException;
-use Yiisoft\Db\Pgsql\Connection;
-use Yiisoft\Db\TestUtility\TestConnectionTrait;
-use Yiisoft\Db\Transaction\Transaction;
+use Yiisoft\Db\TestSupport\TestConnectionTrait;
+use Yiisoft\Db\Transaction\TransactionInterface;
 
 /**
  * @group pgsql
@@ -21,156 +20,85 @@ final class ConnectionTest extends TestCase
 
     public function testConnection(): void
     {
-        $this->assertIsObject($this->getConnection(true));
-    }
-
-    public function testConstruct(): void
-    {
-        $db = $this->getConnection();
-
-        $this->assertEquals(self::DB_DSN, $db->getDsn());
+        $this->assertIsObject($this->getConnection());
     }
 
     public function testGetDriverName(): void
     {
         $db = $this->getConnection();
-
-        $this->assertEquals('pgsql', $db->getDriverName());
+        $this->assertEquals('pgsql', $db->getDriver()->getDriverName());
     }
 
     public function testInitConnection(): void
     {
         $db = $this->getConnection();
-
         $db->setEmulatePrepare(true);
-
         $db->open();
-
         $this->assertTrue($db->getEmulatePrepare());
-
         $db->close();
-    }
-
-    public function testOpenClose(): void
-    {
-        $db = $this->getConnection();
-
-        $this->assertFalse($db->isActive());
-        $this->assertNull($db->getPDO());
-
-        $db->open();
-
-        $this->assertTrue($db->isActive());
-        $this->assertInstanceOf(PDO::class, $db->getPDO());
-
-        $db->close();
-
-        $this->assertFalse($db->isActive());
-        $this->assertNull($db->getPDO());
-
-        $db = $this->createConnection('unknown::memory:');
-
-        $this->expectException(Exception::class);
-        $this->expectExceptionMessage('could not find driver');
-
-        $db->open();
     }
 
     public function testQuoteValue(): void
     {
         $db = $this->getConnection();
+        $quoter = $db->getQuoter();
 
-        $this->assertEquals(123, $db->quoteValue(123));
-        $this->assertEquals("'string'", $db->quoteValue('string'));
-        $this->assertEquals("'It''s interesting'", $db->quoteValue("It's interesting"));
+        $this->assertEquals(123, $quoter->quoteValue(123));
+        $this->assertEquals("'string'", $quoter->quoteValue('string'));
+        $this->assertEquals("'It''s interesting'", $quoter->quoteValue("It's interesting"));
     }
 
     public function testQuoteTableName(): void
     {
         $db = $this->getConnection();
+        $quoter = $db->getQuoter();
 
-        $this->assertEquals('"table"', $db->quoteTableName('table'));
-        $this->assertEquals('"table"', $db->quoteTableName('"table"'));
-        $this->assertEquals('"schema"."table"', $db->quoteTableName('schema.table'));
-        $this->assertEquals('"schema"."table"', $db->quoteTableName('schema."table"'));
-        $this->assertEquals('"schema"."table"', $db->quoteTableName('"schema"."table"'));
-        $this->assertEquals('{{table}}', $db->quoteTableName('{{table}}'));
-        $this->assertEquals('(table)', $db->quoteTableName('(table)'));
+        $this->assertEquals('"table"', $quoter->quoteTableName('table'));
+        $this->assertEquals('"table"', $quoter->quoteTableName('"table"'));
+        $this->assertEquals('"schema"."table"', $quoter->quoteTableName('schema.table'));
+        $this->assertEquals('"schema"."table"', $quoter->quoteTableName('schema."table"'));
+        $this->assertEquals('"schema"."table"', $quoter->quoteTableName('"schema"."table"'));
+        $this->assertEquals('{{table}}', $quoter->quoteTableName('{{table}}'));
+        $this->assertEquals('(table)', $quoter->quoteTableName('(table)'));
     }
 
     public function testQuoteColumnName(): void
     {
         $db = $this->getConnection();
+        $quoter = $db->getQuoter();
 
-        $this->assertEquals('"column"', $db->quoteColumnName('column'));
-        $this->assertEquals('"column"', $db->quoteColumnName('"column"'));
-        $this->assertEquals('[[column]]', $db->quoteColumnName('[[column]]'));
-        $this->assertEquals('{{column}}', $db->quoteColumnName('{{column}}'));
-        $this->assertEquals('(column)', $db->quoteColumnName('(column)'));
+        $this->assertEquals('"column"', $quoter->quoteColumnName('column'));
+        $this->assertEquals('"column"', $quoter->quoteColumnName('"column"'));
+        $this->assertEquals('[[column]]', $quoter->quoteColumnName('[[column]]'));
+        $this->assertEquals('{{column}}', $quoter->quoteColumnName('{{column}}'));
+        $this->assertEquals('(column)', $quoter->quoteColumnName('(column)'));
 
-        $this->assertEquals('"column"', $db->quoteSql('[[column]]'));
-        $this->assertEquals('"column"', $db->quoteSql('{{column}}'));
+        $this->assertEquals('"column"', $quoter->quoteSql('[[column]]'));
+        $this->assertEquals('"column"', $quoter->quoteSql('{{column}}'));
     }
 
     public function testQuoteFullColumnName(): void
     {
         $db = $this->getConnection();
+        $quoter = $db->getQuoter();
 
-        $this->assertEquals('"table"."column"', $db->quoteColumnName('table.column'));
-        $this->assertEquals('"table"."column"', $db->quoteColumnName('table."column"'));
-        $this->assertEquals('"table"."column"', $db->quoteColumnName('"table".column'));
-        $this->assertEquals('"table"."column"', $db->quoteColumnName('"table"."column"'));
+        $this->assertEquals('"table"."column"', $quoter->quoteColumnName('table.column'));
+        $this->assertEquals('"table"."column"', $quoter->quoteColumnName('table."column"'));
+        $this->assertEquals('"table"."column"', $quoter->quoteColumnName('"table".column'));
+        $this->assertEquals('"table"."column"', $quoter->quoteColumnName('"table"."column"'));
 
-        $this->assertEquals('[[table.column]]', $db->quoteColumnName('[[table.column]]'));
-        $this->assertEquals('{{table}}."column"', $db->quoteColumnName('{{table}}.column'));
-        $this->assertEquals('{{table}}."column"', $db->quoteColumnName('{{table}}."column"'));
-        $this->assertEquals('{{table}}.[[column]]', $db->quoteColumnName('{{table}}.[[column]]'));
-        $this->assertEquals('{{%table}}."column"', $db->quoteColumnName('{{%table}}.column'));
-        $this->assertEquals('{{%table}}."column"', $db->quoteColumnName('{{%table}}."column"'));
+        $this->assertEquals('[[table.column]]', $quoter->quoteColumnName('[[table.column]]'));
+        $this->assertEquals('{{table}}."column"', $quoter->quoteColumnName('{{table}}.column'));
+        $this->assertEquals('{{table}}."column"', $quoter->quoteColumnName('{{table}}."column"'));
+        $this->assertEquals('{{table}}.[[column]]', $quoter->quoteColumnName('{{table}}.[[column]]'));
+        $this->assertEquals('{{%table}}."column"', $quoter->quoteColumnName('{{%table}}.column'));
+        $this->assertEquals('{{%table}}."column"', $quoter->quoteColumnName('{{%table}}."column"'));
 
-        $this->assertEquals('"table"."column"', $db->quoteSql('[[table.column]]'));
-        $this->assertEquals('"table"."column"', $db->quoteSql('{{table}}.[[column]]'));
-        $this->assertEquals('"table"."column"', $db->quoteSql('{{table}}."column"'));
-        $this->assertEquals('"table"."column"', $db->quoteSql('{{%table}}.[[column]]'));
-        $this->assertEquals('"table"."column"', $db->quoteSql('{{%table}}."column"'));
-    }
-
-    public function testTransactionIsolation(): void
-    {
-        $db = $this->getConnection(true);
-
-        $transaction = $db->beginTransaction();
-
-        $transaction->setIsolationLevel(Transaction::READ_UNCOMMITTED);
-
-        $transaction->commit();
-
-        $transaction = $db->beginTransaction();
-
-        $transaction->setIsolationLevel(Transaction::READ_COMMITTED);
-
-        $transaction->commit();
-
-        $transaction = $db->beginTransaction();
-
-        $transaction->setIsolationLevel(Transaction::REPEATABLE_READ);
-
-        $transaction->commit();
-
-        $transaction = $db->beginTransaction();
-
-        $transaction->setIsolationLevel(Transaction::SERIALIZABLE);
-
-        $transaction->commit();
-
-        $transaction = $db->beginTransaction();
-
-        $transaction->setIsolationLevel(Transaction::SERIALIZABLE . ' READ ONLY DEFERRABLE');
-
-        $transaction->commit();
-
-        /* should not be any exception so far */
-        $this->assertTrue(true);
+        $this->assertEquals('"table"."column"', $quoter->quoteSql('[[table.column]]'));
+        $this->assertEquals('"table"."column"', $quoter->quoteSql('{{table}}.[[column]]'));
+        $this->assertEquals('"table"."column"', $quoter->quoteSql('{{table}}."column"'));
+        $this->assertEquals('"table"."column"', $quoter->quoteSql('{{%table}}.[[column]]'));
+        $this->assertEquals('"table"."column"', $quoter->quoteSql('{{%table}}."column"'));
     }
 
     /**
@@ -180,21 +108,20 @@ final class ConnectionTest extends TestCase
      */
     public function testGetPdoAfterClose(): void
     {
+        $this->markTestSkipped('Only for master/slave');
+
         $db = $this->getConnection();
 
-        $db->setSlave('1', $this->createConnection(self::DB_DSN));
-
+        $db->setSlave('1', $this->getConnection());
         $this->assertNotNull($db->getSlavePdo(false));
 
         $db->close();
 
         $masterPdo = $db->getMasterPdo();
-
         $this->assertNotFalse($masterPdo);
         $this->assertNotNull($masterPdo);
 
         $slavePdo = $db->getSlavePdo(false);
-
         $this->assertNotFalse($slavePdo);
         $this->assertNotNull($slavePdo);
         $this->assertNotSame($masterPdo, $slavePdo);
@@ -202,37 +129,32 @@ final class ConnectionTest extends TestCase
 
     public function testServerStatusCacheWorks(): void
     {
+        $this->markTestSkipped('Only for master/slave');
+
+        $db = $this->getConnection();
         $cacheKeyNormalizer = new CacheKeyNormalizer();
 
-        $db = $this->getConnection(true);
-
-        $db->setMaster('1', $this->createConnection(self::DB_DSN));
-
+        $db->setMaster('1', $this->getConnection());
         $db->setShuffleMasters(false);
 
         $cacheKey = $cacheKeyNormalizer->normalize(
-            ['Yiisoft\Db\Connection\Connection::openFromPoolSequentially', $db->getDsn()]
+            ['Yiisoft\Db\Connection\Connection::openFromPoolSequentially', $db->getDriver()->getDsn()]
         );
-
         $this->assertFalse($this->cache->psr()->has($cacheKey));
 
         $db->open();
-
         $this->assertFalse(
             $this->cache->psr()->has($cacheKey),
             'Connection was successful – cache must not contain information about this DSN'
         );
 
         $db->close();
-
         $db = $this->getConnection();
-
         $cacheKey = $cacheKeyNormalizer->normalize(
             ['Yiisoft\Db\Connection\Connection::openFromPoolSequentially', 'host:invalid']
         );
 
-        $db->setMaster('1', $this->createConnection('host:invalid'));
-
+        $db->setMaster('1', $this->getConnection(false, 'host:invalid'));
         $db->setShuffleMasters(true);
 
         try {
@@ -250,33 +172,29 @@ final class ConnectionTest extends TestCase
 
     public function testServerStatusCacheCanBeDisabled(): void
     {
-        $cacheKeyNormalizer = new CacheKeyNormalizer();
+        $this->markTestSkipped('Only for master/slave');
 
         $db = $this->getConnection();
+        $cacheKeyNormalizer = new CacheKeyNormalizer();
 
-        $db->setMaster('1', $this->createConnection(self::DB_DSN));
-
+        $db->setMaster('1', $this->getConnection());
         $this->schemaCache->setEnable(false);
-
         $db->setShuffleMasters(false);
 
         $cacheKey = $cacheKeyNormalizer->normalize(
-            ['Yiisoft\Db\Connection\Connection::openFromPoolSequentially', $db->getDsn()]
+            ['Yiisoft\Db\Connection\Connection::openFromPoolSequentially', $db->getDriver()->getDsn()]
         );
-
         $this->assertFalse($this->cache->psr()->has($cacheKey));
 
         $db->open();
-
         $this->assertFalse($this->cache->psr()->has($cacheKey), 'Caching is disabled');
 
         $db->close();
-
         $cacheKey = $cacheKeyNormalizer->normalize(
             ['Yiisoft\Db\Connection\Connection::openFromPoolSequentially', 'host:invalid']
         );
 
-        $db->setMaster('1', $this->createConnection('host:invalid'));
+        $db->setMaster('1', $this->getConnection(false, 'host:invalid'));
 
         try {
             $db->open();
@@ -285,6 +203,65 @@ final class ConnectionTest extends TestCase
 
         $this->assertFalse($this->cache->psr()->has($cacheKey), 'Caching is disabled');
 
+        $db->close();
+    }
+
+    public function testTransactionIsolation(): void
+    {
+        $db = $this->getConnection(true);
+
+        $transaction = $db->beginTransaction();
+        $transaction->setIsolationLevel(TransactionInterface::READ_UNCOMMITTED);
+        $transaction->commit();
+        $transaction = $db->beginTransaction();
+        $transaction->setIsolationLevel(TransactionInterface::READ_COMMITTED);
+        $transaction->commit();
+        $transaction = $db->beginTransaction();
+        $transaction->setIsolationLevel(TransactionInterface::REPEATABLE_READ);
+        $transaction->commit();
+        $transaction = $db->beginTransaction();
+        $transaction->setIsolationLevel(TransactionInterface::SERIALIZABLE);
+        $transaction->commit();
+        $transaction = $db->beginTransaction();
+        $transaction->setIsolationLevel(TransactionInterface::SERIALIZABLE . ' READ ONLY DEFERRABLE');
+        $transaction->commit();
+
+        /* should not be any exception so far */
+        $this->assertTrue(true);
+    }
+
+    public function testTransactionShortcutCustom(): void
+    {
+        $db = $this->getConnection(true);
+
+        $result = $db->transaction(static function (ConnectionInterface $db) {
+            $db->createCommand()->insert('profile', ['description' => 'test transaction shortcut'])->execute();
+            return true;
+        }, TransactionInterface::READ_UNCOMMITTED);
+
+        $this->assertTrue($result, 'transaction shortcut valid value should be returned from callback');
+
+        $profilesCount = $db->createCommand(
+            "SELECT COUNT(*) FROM profile WHERE description = 'test transaction shortcut';"
+        )->queryScalar();
+
+        $this->assertEquals(1, $profilesCount, 'profile should be inserted in transaction shortcut');
+    }
+
+    public function testSettingDefaultAttributes(): void
+    {
+        $db = $this->getConnection();
+        $this->assertEquals(PDO::ERRMODE_EXCEPTION, $db->getActivePDO()->getAttribute(PDO::ATTR_ERRMODE));
+        $db->close();
+
+        $db->setEmulatePrepare(true);
+        $db->open();
+        $this->assertEquals(true, $db->getActivePDO()->getAttribute(PDO::ATTR_EMULATE_PREPARES));
+        $db->close();
+
+        $db->setEmulatePrepare(false);
+        $db->open();
+        $this->assertEquals(false, $db->getActivePDO()->getAttribute(PDO::ATTR_EMULATE_PREPARES));
         $db->close();
     }
 }
