@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Yiisoft\Db\Pgsql\Tests;
 
 use Yiisoft\Db\Query\Query;
-use Yiisoft\Db\TestUtility\TestQueryTrait;
+use Yiisoft\Db\TestSupport\TestQueryTrait;
 
 /**
  * @group pgsql
@@ -29,25 +29,25 @@ final class QueryTest extends TestCase
             ]
         )->execute();
 
-        $this->assertEquals(1, (new Query($db))->from('bool_values')->where('bool_col = TRUE')->count('*', $db));
-        $this->assertEquals(1, (new Query($db))->from('bool_values')->where('bool_col = FALSE')->count('*', $db));
+        $this->assertEquals(1, (new Query($db))->from('bool_values')->where('bool_col = TRUE')->count('*'));
+        $this->assertEquals(1, (new Query($db))->from('bool_values')->where('bool_col = FALSE')->count('*'));
         $this->assertEquals(
             2,
-            (new Query($db))->from('bool_values')->where('bool_col IN (TRUE, FALSE)')->count('*', $db)
+            (new Query($db))->from('bool_values')->where('bool_col IN (TRUE, FALSE)')->count('*')
         );
-        $this->assertEquals(1, (new Query($db))->from('bool_values')->where(['bool_col' => true])->count('*', $db));
-        $this->assertEquals(1, (new Query($db))->from('bool_values')->where(['bool_col' => false])->count('*', $db));
+        $this->assertEquals(1, (new Query($db))->from('bool_values')->where(['bool_col' => true])->count('*'));
+        $this->assertEquals(1, (new Query($db))->from('bool_values')->where(['bool_col' => false])->count('*'));
         $this->assertEquals(
             2,
-            (new Query($db))->from('bool_values')->where(['bool_col' => [true, false]])->count('*', $db)
+            (new Query($db))->from('bool_values')->where(['bool_col' => [true, false]])->count('*')
         );
         $this->assertEquals(
             1,
-            (new Query($db))->from('bool_values')->where('bool_col = :bool_col', ['bool_col' => true])->count('*', $db)
+            (new Query($db))->from('bool_values')->where('bool_col = :bool_col', ['bool_col' => true])->count('*')
         );
         $this->assertEquals(
             1,
-            (new Query($db))->from('bool_values')->where('bool_col = :bool_col', ['bool_col' => false])->count('*', $db)
+            (new Query($db))->from('bool_values')->where('bool_col = :bool_col', ['bool_col' => false])->count('*')
         );
     }
 }
